+++ bugsnag-android-core/src/main/java/com/bugsnag/android/NativeInterface.java
@@ -445,4 +445,20 @@ public class NativeInterface {
     public static Session getCurrentSession() {
         return getClient().sessionTracker.getCurrentSession();
     }
+
+    /**
+     * Marks the launch period as complete
+     */
+    public static void markLaunchCompleted() {
+        getClient().markLaunchCompleted();
+    }
+
+    /**
+     * Get the last run info object
+     */
+    @Nullable
+    public static LastRunInfo getLastRunInfo() {
+        return getClient().getLastRunInfo();
+    }
+
 }